make mocha test resuable
