+++ tests/testing.ts
@@ -1,6 +1,7 @@
 import path from 'path';
 import { CompileFailedError } from 'solc-typed-ast';
-import { findAllFiles, findCairoSourceFilePaths } from '../src/io';
+import { describe } from 'mocha';
+import { findAllFiles, findCairoSourceFilePaths, findSolSourceFilePaths } from '../src/io';
 import { compileSolFiles } from '../src/solCompile';
 import { compileCairo1 } from '../src/starknetCli';
 import { transpile } from '../src/transpiler';
@@ -13,6 +14,8 @@ import { groupBy, printCompileErrors } from '../src/utils/utils';
 import * as fs from 'fs';
 import { outputFileSync } from '../src/utils/fs';
 import { error } from '../src/utils/formatting';
+
+import { expect } from 'chai';
 
 export const solCompileResultTypes = [
   'SolCompileFailed',
@@ -40,6 +43,8 @@ const ResultTypeOrder = [
   'WillNotSupport',
   'SolCompileFailed',
 ];
+
+const TIME_LIMIT = 2 * 60 * 60 * 1000;
 
 export function preTestChecks(warpTestFolder: string): boolean {
   if (!checkNoCairo(warpTestFolder) || !checkNoJson(warpTestFolder)) {
@@ -216,4 +221,81 @@ function deleteJson(path: string): void {
   findAllFiles(path, true)
     .filter((file) => file.endsWith('.json'))
     .forEach((file) => fs.unlinkSync(file));
+}
+
+export function runTests(
+  expectedResults: Map<string, ResultType>,
+  warpTest: string,
+  warpTestFolder: string,
+  warpCompilationTestPath: string,
+  contractsFolder: string,
+) {
+  describe('Running compilation tests', function () {
+    this.timeout(TIME_LIMIT);
+
+    let onlyResults: boolean, unsafe: boolean, force: boolean, exact: boolean;
+    let filter: string | undefined;
+
+    const results = new Map<string, ResultType>();
+
+    this.beforeAll(() => {
+      onlyResults = process.argv.includes('--only-results') ? true : false;
+      unsafe = process.argv.includes('--unsafe') ? true : false;
+      force = process.argv.includes('--force') ? true : false;
+      exact = process.argv.includes('--exact') ? true : false;
+      if (force) {
+        postTestCleanup(warpTestFolder);
+      } else {
+        if (!preTestChecks(warpTestFolder)) return;
+      }
+      filter = process.env.FILTER;
+    });
+
+    describe(`Running warp compilation tests on ${contractsFolder} solidity files`, async function () {
+      findSolSourceFilePaths(warpCompilationTestPath, true).forEach((file) => {
+        if (filter === undefined || file.includes(filter)) {
+          let compileResult: { result: ResultType; cairoFiles?: string[] };
+          const expectedResult: ResultType | undefined = expectedResults.get(
+            path.join(warpTest, file),
+          );
+
+          describe(`Running compilation test on ${file}`, async function () {
+            it(`Running warp compile on ${file}`, async () => {
+              compileResult = runSolFileTest(warpTest, file, results, onlyResults, unsafe);
+              expect(expectedResult).to.not.be.undefined;
+              if (expectedResult === 'Success') {
+                expect(compileResult.result).to.equal('Success');
+              }
+              if (expectedResult !== undefined && solCompileResultTypes.includes(expectedResult)) {
+                expect(compileResult.result).to.equal(expectedResult);
+              }
+            });
+            if (expectedResult !== undefined && cairoCompileResultTypes.includes(expectedResult)) {
+              it(`Running cairo compile on ${file}`, async () => {
+                if (compileResult.cairoFiles !== undefined) {
+                  compileResult.cairoFiles.forEach((cairoFile) => {
+                    const cairoCompileResult = runCairoFileTest(cairoFile, results, onlyResults);
+                    expect(cairoCompileResult).to.equal(expectedResult);
+                  });
+                }
+              });
+            }
+          });
+        }
+      });
+    });
+
+    this.afterAll(() => {
+      const testsWithUnexpectedResults = getTestsWithUnexpectedResults(expectedResults, results);
+      printResults(expectedResults, results, testsWithUnexpectedResults);
+      postTestCleanup(warpTestFolder);
+      if (exact) {
+        if (testsWithUnexpectedResults.length > 0) {
+          throw new Error(
+            error(`${testsWithUnexpectedResults.length} test(s) had unexpected outcome(s)`),
+          );
+        }
+      }
+    });
+  });
 }
